Add certain controls while loading
Fixes #78.
