+++ components/loading-view.js
@@ -1,8 +1,8 @@
 const React = require('react');
 const cn = require('classnames');
 const ReactTooltip = require('react-tooltip');
-const sendToAddon = require('../lib/send-to-addon');
 const sendMetricsEvent = require('../lib/send-metrics-event.js');
+const GeneralControls = require('./general-controls.js');
 
 module.exports = React.createClass({
   getInitialState: function() {
@@ -14,18 +14,16 @@ module.exports = React.createClass({
   leaveView: function() {
     this.setState({hovered: false});
   },
-  close: function() {
-    sendMetricsEvent('loading_view', 'close');
-    sendToAddon({action: 'close'});
-  },
   render: function() {
     sendMetricsEvent('loading_view', 'render');
     return (
-        <div className={'loading'} onMouseEnter={this.enterView} onMouseLeave={this.leaveView}>
+        <div className='loading' onMouseEnter={this.enterView} onMouseLeave={this.leaveView}>
           <ReactTooltip place='bottom' effect='solid' />
-
-          <a className={cn('close', {hidden: this.state.hovered})}
-             onClick={this.close} data-tip='Close' />
+          
+          <div className={cn('controls', {hidden: !this.state.hovered, minimized: this.props.minimized})}>
+            <div className='left' />
+            <GeneralControls props={this.props} />
+          </div>       
 
           <img src={'img/loading-bars.svg'} alt={'loading animation'}
                width={64} height={64}></img>
